requirements.txt
crewai
crewai_tools
python-dotenv
openai
pydub
agentops
streamlit